# tmp/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusCore API",
    "description": "NexusCore API - FastAPI implementation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/execute": {
      "post": {
        "tags": [
          "execute"
        ],
        "summary": "Run self-healing job",
        "description": "Execute エンドポイント\n\n指定されたプロジェクトパスに対して self-healing 実行をトリガーし、\n実行IDおよび現在のステータスを返します。\n\nこのエンドポイントは認証済みクライアントからのみ利用可能です。\n（CR-FASTAPI-003 で認証を追加予定）\n\nArgs:\n    payload: 実行リクエスト（requirement, project_path, constitution_text）\n\nReturns:\n    ExecuteResponse: タスクIDとステータスURLを含むレスポンス\n\nRaises:\n    HTTPException: バリデーションエラーまたは内部エラー時",
        "operationId": "execute_endpoint_api_v1_execute_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "API Key for authentication",
              "title": "X-Api-Key"
            },
            "description": "API Key for authentication"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/status/{task_id}": {
      "get": {
        "tags": [
          "execute"
        ],
        "summary": "Get task status",
        "description": "Get Task Status エンドポイント\n\n指定されたタスクIDの現在の状態を返します。\n既存の Flask 実装 (`get_task_status`) と互換性を保つ。\n\nArgs:\n    task_id: タスクID（UUID形式）\n\nReturns:\n    ExecuteStatusResponse: タスクの状態情報\n\nRaises:\n    HTTPException: タスクが見つからない場合（404）",
        "operationId": "get_task_status_api_v1_status__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "API Key for authentication",
              "title": "X-Api-Key"
            },
            "description": "API Key for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecuteStatusResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/github/webhook": {
      "post": {
        "tags": [
          "github",
          "webhook"
        ],
        "summary": "GitHub Webhook endpoint",
        "description": "GitHub Webhook エンドポイント\n\nGitHub pull_request イベントを受信して Self-Healing Service を実行する。\n既存の Flask 実装 (`handle_github_webhook`) と互換性を保つ。\n\nArgs:\n    request: FastAPI Request オブジェクト（ボディ取得用）\n    x_github_event: X-GitHub-Event ヘッダー\n    x_github_delivery: X-GitHub-Delivery ヘッダー\n    x_hub_signature_256: X-Hub-Signature-256 ヘッダー（署名検証用）\n\nReturns:\n    GitHubWebhookResponse: Webhook処理結果\n\nRaises:\n    HTTPException: 署名検証失敗時（401）または内部エラー時（500）",
        "operationId": "github_webhook_endpoint_api_v1_github_webhook_post",
        "parameters": [
          {
            "name": "X-GitHub-Event",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Github-Event"
            }
          },
          {
            "name": "X-GitHub-Delivery",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Github-Delivery"
            }
          },
          {
            "name": "X-Hub-Signature-256",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Hub-Signature-256"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GitHubWebhookResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "エラーコード"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "人間が読めるエラーメッセージ"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "エラー詳細モデル\n\nAttributes:\n    code: エラーコード（例: \"PROJECT_NOT_FOUND\", \"VALIDATION_ERROR\"）\n    message: 人間が読めるエラーメッセージ"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "エラー詳細情報"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "標準化されたエラーレスポンスモデル\n\nすべての API エンドポイントで統一されたエラー構造。\n\nAttributes:\n    error: エラー詳細情報",
        "example": {
          "error": {
            "code": "PROJECT_NOT_FOUND",
            "message": "Project with id 123 not found"
          }
        }
      },
      "ExecuteRequest": {
        "properties": {
          "requirement": {
            "type": "string",
            "minLength": 1,
            "title": "Requirement",
            "description": "実行する要件"
          },
          "project_path": {
            "type": "string",
            "minLength": 1,
            "title": "Project Path",
            "description": "プロジェクトのパス"
          },
          "constitution_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Constitution Text",
            "description": "プロジェクト憲法のテキスト（デフォルト: 'Default constitution.'）"
          }
        },
        "type": "object",
        "required": [
          "requirement",
          "project_path"
        ],
        "title": "ExecuteRequest",
        "description": "Execute リクエストモデル\n\nAttributes:\n    requirement: 実行する要件（必須）\n    project_path: プロジェクトのパス（必須）\n    constitution_text: プロジェクト憲法のテキスト（任意）",
        "example": {
          "constitution_text": "Write clean, maintainable code.",
          "project_path": "/path/to/project",
          "requirement": "Add a new feature"
        }
      },
      "ExecuteResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "タスク受け入れメッセージ"
          },
          "task_id": {
            "type": "string",
            "title": "Task Id",
            "description": "タスクID（UUID形式）"
          },
          "status_url": {
            "type": "string",
            "title": "Status Url",
            "description": "ステータス確認用URL（相対パス）"
          }
        },
        "type": "object",
        "required": [
          "message",
          "task_id",
          "status_url"
        ],
        "title": "ExecuteResponse",
        "description": "Execute レスポンスモデル\n\nAttributes:\n    message: メッセージ\n    task_id: タスクID（UUID形式）\n    status_url: ステータス確認用URL",
        "example": {
          "message": "Task accepted and is running in the background.",
          "status_url": "/api/v1/status/123e4567-e89b-12d3-a456-426614174000",
          "task_id": "123e4567-e89b-12d3-a456-426614174000"
        }
      },
      "ExecuteStatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "タスクの状態"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "ステータスメッセージ"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "ExecuteStatusResponse",
        "description": "Execute Status レスポンスモデル\n\nタスクの状態を表すモデル。既存のFlask実装では `tasks` 辞書の値がそのまま返されるため、\n柔軟な構造を許容する必要がある。\n\nAttributes:\n    status: タスクの状態（\"running\", \"completed\", \"error\" など）\n    message: ステータスメッセージ"
      },
      "GitHubWebhookResponse": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "description": "Webhookが受け入れられたかどうか"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result",
            "description": "実行結果"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason",
            "description": "拒否理由"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "エラーメッセージ"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "skipped",
                  "fixed",
                  "not_fixed",
                  "no_issues",
                  "error"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "ステータス"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary",
            "description": "サマリー"
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "GitHubWebhookResponse",
        "description": "GitHub Webhook レスポンスモデル\n\n既存のFlask実装のレスポンス形式に準拠。\n\nAttributes:\n    accepted: Webhookが受け入れられたかどうか\n    result: 実行結果（accepted=Trueの場合）\n    reason: 拒否理由（accepted=Falseの場合）\n    error: エラーメッセージ（エラー時）\n    status: ステータス（\"skipped\", \"fixed\", \"not_fixed\", \"no_issues\", \"error\" など）\n    summary: サマリー",
        "example": {
          "accepted": true,
          "result": {
            "run_id": "sh-123",
            "status": "fixed",
            "summary": "Tests are now passing"
          }
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}